+++ Locket/domain/src/main/java/com/ssafy/locket/model/payment_history/PaymentCalendarItem.kt
@@ -2,9 +2,10 @@ package com.ssafy.locket.model.payment_history
 
 import com.ssafy.locket.model.base.BaseModel
 import kotlinx.parcelize.Parcelize
+import java.math.BigDecimal
 
 @Parcelize
 data class PaymentCalendarItem (
-    val amount: Int,
+    val amount: BigDecimal,
     val date: String
 ): BaseModel
